RFC: move functions that are not reused to their dedicated modules
